[UPDATE] api.py code review
